Rename htables to parasha
