+++ hdate/parasha.py
